Moves Block component to shared
